LPS-55159 Move ProjectDataUtil to package com.liferay.cobertura.coveragedata
